requirements.txt
altair==6.0.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
fastapi==0.115.6
Flask==3.1.3
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
narwhals==2.16.0
networkx==3.6.1
numpy==2.4.2
osmium==4.3.0
packaging==26.0
pandas==2.3.3
pillow==12.1.0
plotly==6.4.0
protobuf==6.33.5
python-dotenv==1.0.1
pyarrow==23.0.0
pydeck==0.9.1
pywebview==5.4
pynput==1.8.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
smmap==5.0.2
streamlit==1.54.0
streamlit-searchbox==0.1.24
tenacity==9.1.3
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn[standard]==0.32.1
watchdog==6.0.0
